Avoid use of `__file__` in scribus part
With the Scribus build of Ubuntu 20.04, this apparently fails with a
NameError claiming that `__file__` is not defined.

This change simply removes the code because the converter does not do
any local imports ATM.
In the future, we can see if there is some other way of supporting local
imports, e. g. passing in the path as argument.